refactor basetypechecker to have compositechecker supertype
